Upgrade to OH3

+++ openhab/mounts/etc/cont-init.d/00-jython.sh
@@ -1,6 +1,13 @@
 #!/bin/sh
+
 AUTOMATION_LIB_PYTHON_PATH=${OPENHAB_HOME}/conf/automation/lib/python
 AUTOMATION_JSR223_PYTHON_PATH=${OPENHAB_HOME}/conf/automation/jsr223/python
-find $AUTOMATION_LIB_PYTHON_PATH -type f -name '*\$py.class' -delete
-find "$AUTOMATION_LIB_PYTHON_PATH/personal" -type f -name 'core_*.py' -delete
-find $AUTOMATION_JSR223_PYTHON_PATH/personal -type f -name 'core_*.py' -delete
+if [ -d ${AUTOMATION_LIB_PYTHON_PATH} ]; then
+    find $AUTOMATION_LIB_PYTHON_PATH -type f -name '*\$py.class' -delete
+    find $AUTOMATION_LIB_PYTHON_PATH -type f -name '*\core_*.py' -delete
+fi
+
+if [ -d ${AUTOMATION_JSR223_PYTHON_PATH} ]; then
+    find $AUTOMATION_JSR223_PYTHON_PATH -type f -name '*\$py.class' -delete
+    find $AUTOMATION_JSR223_PYTHON_PATH -type f -name '*\core_*.py' -delete
+fi
